convert to ts
